# solution/gateway-service/swagger/swaggerDocumentation.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Gateway Service API",
    "version": "1.0.0",
    "description": "This is the API documentation for the Gateway Service."
  },
  "servers": [
    {
      "url": "http://localhost:3003",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/post/creator/id": {
      "get": {
        "summary": "Get content by creator ID",
        "description": "Retrieves all posts and comments created by a specific person along with enriched data (e.g., forum title or parent post data).",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "The creator's person ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "The page number for pagination."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "The maximum number of results per page."
          }
        ],
        "responses": {
          "200": {
            "description": "A list of content items created by the specified creator. Each item may be either a comment or a post.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatorContent"
                }
              }
            }
          },
          "404": {
            "description": "No content found for the specified user ID"
          },
          "500": {
            "description": "An error occurred while loading content by user ID."
          }
        }
      }
    },
    "/post/byOrganization/{type}/{orgId}": {
      "get": {
        "summary": "Get content by organization",
        "description": "Retrieves people associated with a specific organization and for each person, fetches their last post.",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "description": "The organization type.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "University",
                "Company"
              ]
            }
          },
          {
            "name": "orgId",
            "in": "path",
            "description": "The organization ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "The page number for pagination."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "The maximum number of results per page."
          }
        ],
        "responses": {
          "200": {
            "description": "A list containing posts from individuals affiliated with the specified organization, either through employment or academic enrollment.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CreatorPostsContent"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No person found for the given organization."
          },
          "500": {
            "description": "An error occurred while loading post by people's oragnization."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 10
          },
          "total": {
            "type": "integer",
            "example": 50
          },
          "totalPages": {
            "type": "integer",
            "example": 5
          },
          "hasNextPage": {
            "type": "boolean",
            "example": true
          },
          "hasPrevPage": {
            "type": "boolean",
            "example": false
          },
          "hasSearched": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
      "ParentPost": {
        "type": "object",
        "description": "Metadata about the original post that a comment is replying to.",
        "properties": {
          "originalId": {
            "type": "number",
            "description": "Original ID of the parent content.",
            "example": 1168231107462
          },
          "originalType": {
            "type": "string",
            "description": "Type of the parent content (Post or Comment).",
            "example": "Post"
          },
          "id": {
            "type": "number",
            "description": "ID of the parent content.",
            "example": 1168231107462
          },
          "content": {
            "type": "string",
            "description": "Content of the parent post.",
            "example": "About Hannibal"
          }
        }
      },
      "CommentContent": {
        "type": "object",
        "description": "Structure representing a comment or a post with metadata.",
        "properties": {
          "id": {
            "type": "number",
            "description": "Unique ID of the content (post or comment).",
            "example": 1168231107465
          },
          "type": {
            "type": "string",
            "description": "Type of content: 'Post' or 'Comment'.",
            "example": "Comment"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time",
            "description": "Date the content was created.",
            "example": "2012-11-19T19:51:30.896Z"
          },
          "content": {
            "type": "string",
            "description": "Textual content.",
            "example": "thx"
          },
          "forumTitle": {
            "type": "string",
            "description": "Title of the forum where the content was posted.",
            "example": "Album 1 of Jun Li",
            "mandatory": false
          },
          "parentPost": {
            "$ref": "#/components/schemas/ParentPost"
          }
        }
      },
      "CreatorContent": {
        "type": "object",
        "description": "Data structure combining creator information with all content items authored by the person.",
        "properties": {
          "creator": {
            "type": "object",
            "description": "Information about the creator of the content.",
            "properties": {
              "firstName": {
                "type": "string",
                "description": "The creator's first name."
              },
              "lastName": {
                "type": "string",
                "description": "The creator's last name."
              }
            },
            "required": [
              "firstName",
              "lastName"
            ]
          },
          "content": {
            "type": "array",
            "description": "List of posts or comments created by the person, including parent post data where applicable.",
            "items": {
              "$ref": "#/components/schemas/CommentContent"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          },
          "hasSearched": {
            "type": "boolean",
            "description": "Indicates if a search was performed."
          }
        }
      },
      "Post": {
        "type": "object",
        "description": "Detailed information about a post created by a user.",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Internal MongoDB ID of the post.",
            "example": "681252856b3cb58f077f9900"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of when the post was created.",
            "example": "2012-11-08T17:17:38.086Z"
          },
          "id": {
            "type": "number",
            "description": "LDBC-generated unique ID for the post.",
            "example": 1099511630536
          },
          "imageFile": {
            "type": "string",
            "description": "Name of the image file attached to the post.",
            "example": "photo1099511630536.jpg"
          },
          "locationIP": {
            "type": "string",
            "description": "IP address used at the time of posting.",
            "example": "1.29.220.142"
          },
          "browserUsed": {
            "type": "string",
            "description": "Browser used to create the post.",
            "example": "Chrome"
          },
          "length": {
            "type": "number",
            "description": "Length of the post content.",
            "example": 0
          },
          "CreatorPersonId": {
            "type": "number",
            "description": "ID of the person who created the post.",
            "example": 21990232555527
          },
          "ContainerForumId": {
            "type": "number",
            "description": "ID of the forum container for the post.",
            "example": 1099511628047
          },
          "LocationCountryId": {
            "type": "number",
            "description": "ID of the country location associated with the post.",
            "example": 1
          },
          "forumTitle": {
            "type": "string",
            "description": "Title of the forum where the post was made.",
            "example": "Album 1 of Jun Li"
          }
        }
      },
      "CreatorPostsContent": {
        "type": "object",
        "description": "Information about a content creator and the posts they have published.",
        "properties": {
          "id": {
            "type": "number",
            "description": "Unique identifier of the content creator.",
            "example": 21990232555527
          },
          "name": {
            "type": "string",
            "description": "First name of the creator.",
            "example": "Jun"
          },
          "surname": {
            "type": "string",
            "description": "Last name of the creator.",
            "example": "Li"
          },
          "since": {
            "type": "number",
            "description": "Year when the person joined the organization.",
            "example": 2005
          },
          "posts": {
            "type": "array",
            "description": "List of posts authored by the creator.",
            "items": {
              "$ref": "#/components/schemas/Post"
            }
          },
          "pagination": {
            "type": "object",
            "description": "Pagination information for the list of posts.",
            "items": {
              "$ref": "#/components/schemas/Pagination"
            }
          }
        }
      }
    }
  }
}